Rename logs module to logutils
