+++ src/batcontrol/logutils/__init__.py
@@ -1,2 +1,2 @@
 from .logfilelimiter import LogFileLimiter
-from .logging_config import setup_logging
+from .config import setup_logging
